feat(ui): add JobsStore, use of it in List
